# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "description": "Rating API, uses Glicko2 rating algorithim. feel free to contact me for a permenant api token <br><b>Public Bearer token</b>:",
        "version": "0.1.0",
        "title": "RaterApi",
        "contact": {
            "email": "Blankhussien@gmail.com"
        }
    },
...
    "paths": {
        "/games": {
            "get": {
                "tags": [
                    "game"
                ],
                "summary": "Get all games",
                "description": "Return all games, will return empty body if there was none",
                "operationId": "indexGames",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "League of legends"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "example": "1 hour ago"
                                            },
                                            "links": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object",
                                                    "properties": {
                                                        "rel": {
                                                            "type": "string"
                                                        },
                                                        "href": {
                                                            "type": "string"
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                },
                                "meta": {
                                    "type": "object",
                                    "properties": {
                                        "first_page": {
                                            "type": "string"
                                        },
                                        "etc....": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "links": {
                                    "type": "object",
                                    "properties": {
                                        "first_page": {
                                            "type": "string"
                                        },
                                        "etc....": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "game"
                ],
                "summary": "Create a new game",
                "description": "Create a game so you can arrange matches for your own users",
                "operationId": "addGame",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Game details",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Game"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Game created"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
        "/users": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get all users",
                "description": "Return all users, will return empty body if there was none",
                "operationId": "indexUsers",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "John doe"
                                            },
                                            "links": {
                                                "type": "object",
                                                "example": [
                                                    {
                                                        "rel": "self",
                                                        "href": "/users/john-doe"
                                                    },
                                                    {
                                                        "rel": "games",
                                                        "href": "/users/john-doe/games"
                                                    },
                                                    {
                                                        "rel": "matches",
                                                        "href": "/users/john-doe/matches"
                                                    }
                                                ]
                                            }
                                        }
                                    }
                                },
                                "meta": {
                                    "type": "object",
                                    "properties": {
                                        "first_page": {
                                            "type": "string"
                                        },
                                        "etc....": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "links": {
                                    "type": "object",
                                    "properties": {
                                        "first_page": {
                                            "type": "string"
                                        },
                                        "etc....": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Game": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Pubg"
                }
            }
        }
    }
}